requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-ninja==1.3.0
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.12.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
imageio==2.36.1
isort==5.13.2
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
mypy-extensions==1.0.0
nanodjango==0.9.2
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.1
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-image==0.25.0
scipy==1.15.0
selenium==4.27.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
sympy==1.13.1
tifffile==2024.12.12
torch==2.5.1
torchvision==0.20.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
whitenoise==6.8.2
wsproto==1.2.0
